refactor(dns): split dns module into per-command files; fix CNAME conflict errors
Refresh the Domains v3 OpenAPI spec and use it to fix `dns set`/`dns add`
surfacing an opaque error on a CNAME name-exclusivity conflict, plus a
related fix in `domain purchase`. Also break up the ~1760-line dns/mod.rs
into per-command files (list/add/set/delete) with shared helpers split into
records.rs and conflicts.rs, mirroring the domain/ module's layout for
easier navigation. No behavior change from the reorganization itself.
